# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Köyden Şehre API
  version: "1.0.0"
  description: Farm-to-city marketplace backend API
  contact:
    name: Köyden Şehre
servers:
  - url: http://localhost:8080/api/v1
    description: Development
  - url: https://api.koydensehire.com/api/v1
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /farmer-applications:
    post:
      summary: Submit farmer application
      tags: [Applications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invite_code, full_name, phone, password, business_name, producer_type, city, district, village, bio, product_categories, product_examples, kvkk_accepted, platform_terms_accepted, declares_own_production, declares_accurate_location, declares_not_intermediary]
              properties:
                invite_code: { type: string }
                full_name: { type: string }
                phone: { type: string }
                email: { type: string }
                password: { type: string, minLength: 6 }
                business_name: { type: string }
                producer_type: { type: string }
                city: { type: string }
                district: { type: string }
                village: { type: string }
                bio: { type: string }
                product_categories: { type: array, items: { type: string } }
                product_examples: { type: string }
                production_place_type: { type: string }
                kvkk_accepted: { type: boolean }
                platform_terms_accepted: { type: boolean }
                declares_own_production: { type: boolean }
                declares_accurate_location: { type: boolean }
                declares_not_intermediary: { type: boolean }
      responses:
        '201': { description: Application submitted }
        '400': { description: Validation error }
        '409': { description: Conflict }
...
  /admin/applications:
    get:
      summary: List farmer applications
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [pending, needs_video, approved, rejected] } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        '200': { description: Application list }
...